feat: enhance headless table with row selection, skeleton loading, and modular architecture
- Added row selection functionality with checkboxes, select all, and indeterminate states
- Implemented customizable skeleton loading with configurable row count
- Refactored component structure into modular directories (table-component, table-layout)
- Updated documentation with new features, API reference, and comprehensive examples
- Added SelectionConfig and LoadingOptions types for better configuration
