+++ components/headless/table/components/table-layout/types.ts
@@ -1,0 +1,7 @@
+import React from "react";
+
+export type TableLayoutProps = {
+  filters?: React.ReactNode;
+  table: React.ReactNode;
+  pagination?: React.ReactNode;
+};
